marche niquel pour %d mais plus pour %s et %c

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: mavileo <mavileo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 16:20:40 by mavileo           #+#    #+#             */
-/*   Updated: 2019/12/01 16:47:06 by mavileo          ###   ########.fr       */
+/*   Updated: 2019/12/02 22:02:43 by mavileo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t count;
+	int count;
 
 	count = 0;
 	while (s[count])
